Customize language filter and article column behavior
Added 'mdwiki' as a custom language in the language filter and updated the ArticleColumn to pass the language to the PageTitleWidget. Modified SelectedSourcePage to hide 'en' and 'simple' languages and adjusted language filter logic to only show 'mdwiki' as a source language. Also, passed campaign info to TargetArticle in TranslationController.